Allow more then 2^16 - 1 packages as preselection.
Now it's possible to create more than 65535 packages. The limit is
now set to 2^32 - 1 packages.

Reviewed-on: https://trac.frm2.tum.de/review/3758
Tested-by: Jens Krueger <jens.krueger@frm2.tum.de>

+++ tools/simulator/simapp.h
@@ -63,7 +63,7 @@ private:
 	quint16            m_wSpectrumHeight;
 
 	//! stop after this number of packets
-	quint16            m_dwStopPacket;
+	quint32            m_dwStopPacket;
 
 	//! simulation timer
 	quint16            m_wTimerInterval;
